+++ uni/lib/main.dart
@@ -10,23 +10,21 @@ import 'package:uni/controller/local_storage/app_shared_preferences.dart';
 import 'package:uni/controller/middleware.dart';
 import 'package:uni/controller/on_start_up.dart';
 import 'package:uni/model/app_state.dart';
-import 'package:uni/model/locations_page_model.dart';
-import 'package:uni/model/schedule_page_model.dart';
-import 'package:uni/redux/actions.dart';
+import 'package:uni/view/locations/locations.dart';
 import 'package:uni/redux/reducers.dart';
 import 'package:uni/utils/constants.dart' as constants;
-import 'package:uni/view/Pages/about_page_view.dart';
-import 'package:uni/view/Pages/bug_report_page_view.dart';
-import 'package:uni/view/Pages/bus_stop_next_arrivals_page.dart';
-import 'package:uni/view/Pages/calendar_page_view.dart';
-import 'package:uni/view/Pages/course_units_page_view.dart';
-import 'package:uni/view/Pages/exams_page_view.dart';
-import 'package:uni/view/Pages/home_page_view.dart';
-import 'package:uni/view/Pages/logout_route.dart';
-import 'package:uni/view/Pages/splash_page_view.dart';
-import 'package:uni/view/Pages/useful_contacts_card_page_view.dart';
-import 'package:uni/view/Pages/useful_links_card_page_view.dart';
-import 'package:uni/view/Widgets/page_transition.dart';
+import 'package:uni/view/about/about.dart';
+import 'package:uni/view/bug_report/bug_report.dart';
+import 'package:uni/view/bus_stop_next_arrivals/bus_stop_next_arrivals.dart';
+import 'package:uni/view/exams/exams.dart';
+import 'package:uni/view/home/home.dart';
+import 'package:uni/view/calendar/calendar.dart';
+import 'package:uni/view/course_units/course_units.dart';
+import 'package:uni/view/logout_route.dart';
+import 'package:uni/view/splash/splash.dart';
+import 'package:uni/view/useful_info/useful_info.dart';
+import 'package:uni/view/schedule/schedule.dart';
+import 'package:uni/view/common_widgets/page_transition.dart';
 import 'package:uni/view/navigation_service.dart';
 import 'package:uni/view/theme.dart';
 import 'package:uni/view/theme_notifier.dart';
@@ -117,13 +115,9 @@ class MyAppState extends State<MyApp> {
                   case '/${constants.navCalendar}':
                     return PageTransition.makePageTransition(
                         page: const CalendarPageView(), settings: settings);
-                  case '/${constants.navUsefulContacts}':
+                  case '/${constants.navUsefulInfo}':
                     return PageTransition.makePageTransition(
-                        page: const UsefulContactsCardView(),
-                        settings: settings);
-                  case '/${constants.navUsefulLinks}':
-                    return PageTransition.makePageTransition(
-                        page: const UsefulLinksCardView(), settings: settings);
+                        page: const UsefulInfoPageView(), settings: settings);
                   case '/${constants.navAbout}':
                     return PageTransition.makePageTransition(
                         page: const AboutPageView(), settings: settings);
@@ -138,12 +132,5 @@ class MyAppState extends State<MyApp> {
                 return null;
               }),
         ));
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    Timer.periodic(const Duration(seconds: 60),
-        (Timer t) => state.dispatch(SetCurrentTimeAction(DateTime.now())));
   }
 }