+++ uni/lib/model/home_page_model.dart
@@ -1,1 +1,0 @@
-enum FavoriteWidgetType { exams, schedule, printBalance, account, busStops }