+++ uni/lib/view/useful_info/widgets/other_links_card.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:uni/view/common_widgets/generic_expansion_card.dart';
+import 'package:uni/view/useful_info/widgets/link_button.dart';
+
+/// Manages the 'Current account' section inside the user's page (accessible
+/// through the top-right widget with the user picture)
+class OtherLinksCard extends GenericExpansionCard {
+  const OtherLinksCard({Key? key}) : super(key: key);
+
+  @override
+  Widget buildCardContent(BuildContext context) {
+    return Column(children: const [
+      LinkButton(title: 'Impressão', link: 'https://print.up.pt')
+    ]);
+  }
+
+  @override
+  String getTitle() => 'Outros Links';
+}